+++ src/components/common/infiniteComponent.vue
@@ -1,18 +1,17 @@
 <template>
-    <div>
-      <ul ref="listEl">
-        <li v-for="comment in commentsList">
-          {{ comment }}
-        </li>
-      </ul>
-    </div>
+  <div>
+    <ul ref="listEl">
+      <li v-for="comment in commentsList">
+        {{ comment }}
+      </li>
+    </ul>
+  </div>
 </template>
 
 <script setup>
+import { axios } from 'axios';
 
-import { axios } from "axios";
-
- const getComments = async(max, omit) => {
+const getComments = async (max, omit) => {
   try {
     const comments = await axios.get(
       `https://jsonplaceholder.typicode.com/comments?_limit=${max}&_start=${omit}`
@@ -21,5 +20,6 @@ import { axios } from "axios";
   } catch (error) {
     console.error(error);
   }
-}
+};
+export { getComments };
 </script>
